regen with latest emitter
