+++ petstore/clients/javascript/src/api/petInsurance/index.ts
@@ -9,11 +9,9 @@ import {
   Insurance,
   insuranceDeserializer,
   petStoreErrorDeserializer,
-} from "../../models/models.js";
-import {
   InsuranceUpdate,
   insuranceUpdateSerializer,
-} from "../../models/typeSpec/rest/resource/models.js";
+} from "../../models/models.js";
 import {
   StreamableMethod,
   PathUncheckedResponse,
