+++ petstore/clients/javascript/src/api/toys/index.ts
@@ -9,11 +9,9 @@ import {
   petStoreErrorDeserializer,
   Toy,
   toyDeserializer,
-} from "../../models/models.js";
-import {
   ToyCollectionWithNextLink,
   toyCollectionWithNextLinkDeserializer,
-} from "../../models/typeSpec/rest/resource/models.js";
+} from "../../models/models.js";
 import {
   StreamableMethod,
   PathUncheckedResponse,
