+++ petstore/clients/javascript/src/models/index.ts
@@ -3,8 +3,18 @@
 export {
   Insurance,
   PetStoreError,
+  InsuranceUpdate,
+  CheckupUpdate,
   Checkup,
+  CheckupCollectionWithNextLink,
   Owner,
+  OwnerUpdate,
+  OwnerCreate,
+  OwnerCollectionWithNextLink,
   Toy,
+  ToyCollectionWithNextLink,
   Pet,
+  PetUpdate,
+  PetCreate,
+  PetCollectionWithNextLink,
 } from "./models.js";
