--- petstore/clients/javascript/src/models/typeSpec/rest/resource/index.ts
@@ -1,14 +1,0 @@
-// Licensed under the MIT License.
-
-export {
-  InsuranceUpdate,
-  CheckupUpdate,
-  CheckupCollectionWithNextLink,
-  OwnerUpdate,
-  OwnerCreate,
-  OwnerCollectionWithNextLink,
-  ToyCollectionWithNextLink,
-  PetUpdate,
-  PetCreate,
-  PetCollectionWithNextLink,
-} from "./models.js";
