+++ todoApp/clients/javascript/package.json
@@ -11,11 +11,13 @@
     "exports": {
       "./package.json": "./package.json",
       ".": "./src/index.ts",
-      "./models": "./src/models/index.ts",
       "./api": "./src/api/index.ts",
       "./api/todoItems/attachments": "./src/api/todoItems/attachments/index.ts",
       "./api/todoItems": "./src/api/todoItems/index.ts",
-      "./api/users": "./src/api/users/index.ts"
+      "./api/users": "./src/api/users/index.ts",
+      "./models": "./src/models/index.ts",
+      "./models/todoItems": "./src/models/todoItems/index.ts",
+      "./models/users": "./src/models/users/index.ts"
     },
     "dialects": [
       "esm",
@@ -67,24 +69,6 @@
       "require": {
         "types": "./dist/commonjs/index.d.ts",
         "default": "./dist/commonjs/index.js"
-      }
-    },
-    "./models": {
-      "browser": {
-        "types": "./dist/browser/models/index.d.ts",
-        "default": "./dist/browser/models/index.js"
-      },
-      "react-native": {
-        "types": "./dist/react-native/models/index.d.ts",
-        "default": "./dist/react-native/models/index.js"
-      },
-      "import": {
-        "types": "./dist/esm/models/index.d.ts",
-        "default": "./dist/esm/models/index.js"
-      },
-      "require": {
-        "types": "./dist/commonjs/models/index.d.ts",
-        "default": "./dist/commonjs/models/index.js"
       }
     },
     "./api": {
@@ -157,6 +141,60 @@
       "require": {
         "types": "./dist/commonjs/api/users/index.d.ts",
         "default": "./dist/commonjs/api/users/index.js"
+      }
+    },
+    "./models": {
+      "browser": {
+        "types": "./dist/browser/models/index.d.ts",
+        "default": "./dist/browser/models/index.js"
+      },
+      "react-native": {
+        "types": "./dist/react-native/models/index.d.ts",
+        "default": "./dist/react-native/models/index.js"
+      },
+      "import": {
+        "types": "./dist/esm/models/index.d.ts",
+        "default": "./dist/esm/models/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/models/index.d.ts",
+        "default": "./dist/commonjs/models/index.js"
+      }
+    },
+    "./models/todoItems": {
+      "browser": {
+        "types": "./dist/browser/models/todoItems/index.d.ts",
+        "default": "./dist/browser/models/todoItems/index.js"
+      },
+      "react-native": {
+        "types": "./dist/react-native/models/todoItems/index.d.ts",
+        "default": "./dist/react-native/models/todoItems/index.js"
+      },
+      "import": {
+        "types": "./dist/esm/models/todoItems/index.d.ts",
+        "default": "./dist/esm/models/todoItems/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/models/todoItems/index.d.ts",
+        "default": "./dist/commonjs/models/todoItems/index.js"
+      }
+    },
+    "./models/users": {
+      "browser": {
+        "types": "./dist/browser/models/users/index.d.ts",
+        "default": "./dist/browser/models/users/index.js"
+      },
+      "react-native": {
+        "types": "./dist/react-native/models/users/index.d.ts",
+        "default": "./dist/react-native/models/users/index.js"
+      },
+      "import": {
+        "types": "./dist/esm/models/users/index.d.ts",
+        "default": "./dist/esm/models/users/index.js"
+      },
+      "require": {
+        "types": "./dist/commonjs/models/users/index.d.ts",
+        "default": "./dist/commonjs/models/users/index.js"
       }
     }
   },
